[enhancement/3] Split entities
