+++ src/Forecast/index.tsx
@@ -13,7 +13,9 @@ import RemoteData from 'frctl/RemoteData'
 import Either from 'frctl/Either'
 
 import { Effects, Dispatch } from 'core'
-import { TempUnits, DayForecast, getFiveDayForecastForCity } from 'api'
+import { getFiveDayForecastForCity } from 'api'
+import TempUnits from 'entities/TempUnits'
+import DayForecast from 'entities/DayForecast'
 import { Error as HttpError } from 'httpBuilder'
 import WeekRow, { SkeletonWeekRow } from './WeekRow'
 import ErrorReport from './ErrorReport'
