+++ src/entities/TempUnits.ts
@@ -1,0 +1,6 @@
+enum TempUnits {
+  Celcius = 'metric',
+  Fahrenheit = 'imperial'
+}
+
+export default TempUnits
